Share the resources-equivalence assertion between test modules
The assert_resources_equivalent helper was duplicated in
test_resources.py and test_stage.py. It now lives in a shared
resources_equivalence module in the test directory, imported by its
fully-qualified name via implicit namespace packages, with the
repository root added to pytest's pythonpath to make that resolvable.
The helper's headless self-test stays in test_resources.py, which
pytest collects, and imports the shared copy.
