+++ pytest.ini
@@ -7,6 +7,12 @@
 testpaths = service/*/test
 addopts = --import-mode=importlib
 
+# Put the repository root on sys.path so test modules can import shared helpers from
+# their own directory by fully-qualified name (e.g.
+# service.SpaceCenter.test.resources_equivalence) via implicit namespace packages.
+# Fully-qualified imports keep the colliding-basename problem away.
+pythonpath = .
+
 # Stream the framework's progress messages (game loading, mod set in use) live to the
 # console. Only the krpctest logger emits at INFO in this codebase, so this stays quiet.
 log_cli = true
